feat: moved remaining reducers in slices folder
