+++ app/components/UI/ApproveTransactionReview/index.js
@@ -94,7 +94,7 @@ import ApproveTransactionHeader from '../ApproveTransactionHeader';
 import VerifyContractDetails from './VerifyContractDetails/VerifyContractDetails';
 import ShowBlockExplorer from './ShowBlockExplorer';
 import { isNetworkRampNativeTokenSupported } from '../Ramp/common/utils';
-import { getRampNetworks } from '../../../reducers/fiatOrders';
+import { getRampNetworks } from '../../../redux/slices/fiatOrders';
 import SkeletonText from '../Ramp/common/components/SkeletonText';
 import InfoModal from '../../../components/UI/Swaps/components/InfoModal';
 import TransactionBlockaidBanner from '../TransactionBlockaidBanner/TransactionBlockaidBanner';
