+++ app/components/UI/CollectibleModal/CollectibleModal.tsx
@@ -14,7 +14,7 @@ import { baseStyles } from '../../../styles/common';
 import ReusableModal from '../ReusableModal';
 import Routes from '../../../constants/navigation/Routes';
 import CollectibleOverview from '../../UI/CollectibleOverview';
-import { collectiblesSelector } from '../../../reducers/collectibles';
+import { collectiblesSelector } from '../../../redux/slices/collectibles';
 import {
   selectDisplayNftMedia,
   selectIsIpfsGatewayEnabled,
