+++ app/components/UI/Notification/TransactionNotification/index.js
@@ -21,7 +21,7 @@ import Device from '../../../../util/device';
 import ElevatedView from 'react-native-elevated-view';
 import { CANCEL_RATE, SPEED_UP_RATE } from '@metamask/transaction-controller';
 import BigNumber from 'bignumber.js';
-import { collectibleContractsSelector } from '../../../../reducers/collectibles';
+import { collectibleContractsSelector } from '../../../../redux/slices/collectibles';
 import { useTheme } from '../../../../util/theme';
 import {
   selectChainId,
