+++ app/components/UI/Ramp/common/Views/OrdersList/OrdersList.tsx
@@ -21,7 +21,7 @@ import Text, {
 } from '../../../../../../component-library/components/Texts/Text';
 
 import { FIAT_ORDER_PROVIDERS } from '../../../../../../constants/on-ramp';
-import { FiatOrder, getOrders } from '../../../../../../reducers/fiatOrders';
+import { FiatOrder, getOrders } from '../../../../../../redux/slices/fiatOrders';
 import { strings } from '../../../../../../../locales/i18n';
 import { useTheme } from '../../../../../../util/theme';
