+++ app/components/UI/Ramp/common/components/OrderDetails.tsx
@@ -20,7 +20,7 @@ import {
   renderFromTokenMinimalUnit,
   toTokenMinimalUnit,
 } from '../../../../../util/number';
-import { FiatOrder, getProviderName } from '../../../../../reducers/fiatOrders';
+import { FiatOrder, getProviderName } from '../../../../../redux/slices/fiatOrders';
 import useBlockExplorer from '../../../Swaps/utils/useBlockExplorer';
 import Spinner from '../../../AnimatedSpinner';
 import useAnalytics from '../hooks/useAnalytics';
