+++ app/components/UI/Ramp/common/components/OrderListItem/OrderListItem.tsx
@@ -7,7 +7,7 @@ import BaseListItem from '../../../../../Base/ListItem';
 import {
   FiatOrder,
   getProviderName,
-} from '../../../../../../reducers/fiatOrders';
+} from '../../../../../../redux/slices/fiatOrders';
 import { strings } from '../../../../../../../locales/i18n';
 import { toDateFormat } from '../../../../../../util/date';
 import { addCurrencySymbol, renderFiat } from '../../../../../../util/number';
