+++ app/components/Views/Collectible/index.js
@@ -10,7 +10,7 @@ import Modal from 'react-native-modal';
 import CollectibleContractInformation from '../../UI/CollectibleContractInformation';
 import { toggleCollectibleContractModal } from '../../../actions/modals';
 import { toLowerCaseEquals } from '../../../util/general';
-import { collectiblesSelector } from '../../../reducers/collectibles';
+import { collectiblesSelector } from '../../../redux/slices/collectibles';
 import { ThemeContext, mockTheme } from '../../../util/theme';
 
 const createStyles = (colors) =>
